backend/requirements.txt
fastapi
uvicorn
Jinja2
polygon-api-client
python-multipart
python-dotenv